# Improve performance of src_analyse_IRSARRF_SRS() with optimization at -O2