These changes improve the running time of our example when using -O2, which is the default compilation flag on CRAN. This is a mild regression for the -O3 flag, but given that hardly anybody will use that, the compromise seems worth it. Part of #372.

Single core performance using 3 slide windows:
 - with -O2 time goes from 13.5s to 10.7s
 - with -03 time goes from 10.1s to 10.7s

Single core performance using 10 slide windows:
 - with -O2 time goes from 40.9s to 32.6s
 - with -03 time goes from 29.1s to 32.5s